yo it finally works but fr this time
