+++ src/test/java/use_case/GetNewsUseCaseTest.java
@@ -1,18 +1,12 @@
-package interface_adapter;
+package use_case;
 
 import entity.*;
 import interface_adapter.GetNews.GetNewsController;
-import interface_adapter.GetNews.GetNewsPresenter;
-import interface_adapter.GetNews.GetNewsState;
-import interface_adapter.GetNews.GetNewsViewModel;
-import org.apiguardian.api.API;
 import org.junit.jupiter.api.*;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
-import use_case.APIAccessInterface;
 import use_case.GetNews.*;
 
-import javax.swing.text.View;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -24,10 +18,9 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class GetNewsTest {
+public class GetNewsUseCaseTest {
     private APIAccessInterface mockAPI;
     private GetNewsOutputBoundary presenter;
-    private GetNewsInteractor interactor;
 
     @BeforeAll
     public void setUpAPI() throws RuntimeException, APIAccessInterface.TickerNotFoundException {
@@ -68,7 +61,6 @@ public class GetNewsTest {
     @BeforeEach
     public void setUpPresenter() {
         presenter = Mockito.mock(GetNewsOutputBoundary.class);
-        interactor = new GetNewsInteractor(presenter, mockAPI);
     }
 
     @AfterEach
@@ -78,6 +70,11 @@ public class GetNewsTest {
 
     @Test
     public void testInfoFetch() {
+//        GetNewsInputData inputData = new GetNewsInputData("AAPL");
+//        GetNewsInteractor interactor = new GetNewsInteractor(presenter, mockAPI);
+//        interactor.execute(inputData);
+
+        GetNewsInteractor interactor = new GetNewsInteractor(presenter, mockAPI);
         GetNewsController controller = new GetNewsController(interactor);
         controller.execute("AAPL");
 
@@ -92,49 +89,21 @@ public class GetNewsTest {
 
         assertEquals(5, response.getNewsItems().size());
         assertEquals("AAPL", response.getTicker());
-        assertEquals("The Cash-Rich Magnificent 7 For The Long Haul", newsItem.get("headline"));
-        assertEquals("company", newsItem.get("category"));
 
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("hh:mm:ss a yyyy-MM-dd");
         LocalDateTime localDateTime = LocalDateTime.parse(newsItem.get("datetime"), formatter);
         String date = localDateTime.format(formatter);
-        assertEquals("12:00:00 a.m. 2023-11-28", date);
-    }
+        assertEquals("2023-11-28", localDateTime.toLocalDate().toString());
 
-    @Test
-    public void testInterfaceAdapters() {
-        // Test-specific setup.
-        GetNewsController controller = new GetNewsController(interactor);
-        controller.execute("AAPL");
-        ArgumentCaptor<GetNewsOutputData> captor = ArgumentCaptor.forClass(GetNewsOutputData.class);
-        Mockito.verify(presenter).prepareSuccessView(captor.capture());
-        Mockito.verify(presenter, Mockito.never()).prepareFailView(any(String.class));
-        GetNewsOutputData response = captor.getValue();
-
-        // Test Interface Adapters
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        GetNewsViewModel getNewsViewModel = new GetNewsViewModel();
-        GetNewsPresenter getNewsPresenter = new GetNewsPresenter(viewManagerModel, getNewsViewModel);
-
-        getNewsPresenter.prepareSuccessView(response);
-
-        // Verify properties of the state
-        GetNewsState state = getNewsViewModel.getState();
-
-//        assertFalse(state.getRenderNewInfo());
-        assertNull(state.getTickerError());
-
-        Map<String, String> newsItem = state.getNewsItems().get(0);
-        assertEquals(5, state.getNewsItems().size());
-        assertEquals("AAPL", state.getTicker());
         assertEquals("The Cash-Rich Magnificent 7 For The Long Haul", newsItem.get("headline"));
         assertEquals("company", newsItem.get("category"));
     }
 
     @Test
     public void testInvalidTicker() throws APIAccessInterface.TickerNotFoundException {
-        GetNewsController controller = new GetNewsController(interactor);
-        controller.execute("SSS");
+        GetNewsInputData inputData = new GetNewsInputData("SSS");
+        GetNewsInteractor interactor = new GetNewsInteractor(presenter, mockAPI);
+        interactor.execute(inputData);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
 
@@ -148,8 +117,9 @@ public class GetNewsTest {
 
     @Test
     public void testNoApiResponse() throws RuntimeException {
-        GetNewsController controller = new GetNewsController(interactor);
-        controller.execute("API does not respond");
+        GetNewsInputData inputData = new GetNewsInputData("API does not respond");
+        GetNewsInteractor interactor = new GetNewsInteractor(presenter, mockAPI);
+        interactor.execute(inputData);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
